+++ docs/src/components.d.ts
@@ -7,12 +7,12 @@
 
 import '@stencil/core';
 
+import '@deckdeckgo/charts';
+import '@deckdeckgo/core';
+import '@deckdeckgo/highlight-code';
+import '@deckdeckgo/inline-editor';
+import '@deckdeckgo/qrcode';
 import '@ionic/core';
-import 'deckdeckgo';
-import 'deckdeckgo-charts';
-import 'deckdeckgo-highlight-code';
-import 'deckdeckgo-inline-editor';
-import 'deckdeckgo-qrcode';
 import 'ionicons';