+++ studio/src/components.d.ts
@@ -7,14 +7,14 @@
 
 import '@stencil/core';
 
+import '@deckdeckgo/core';
+import '@deckdeckgo/highlight-code';
+import '@deckdeckgo/inline-editor';
+import '@deckdeckgo/lazy-img';
+import '@deckdeckgo/qrcode';
+import '@deckdeckgo/remote';
 import '@ionic/core';
 import 'ionicons';
-import 'deckdeckgo';
-import 'deckdeckgo-highlight-code';
-import 'deckdeckgo-inline-editor';
-import 'deckdeckgo-lazy-img';
-import 'deckdeckgo-qrcode';
-import 'deckdeckgo-remote';
 import {
   EventEmitter,
 } from '@stencil/core';