+++ webcomponents/qrcode/CHANGELOG.md
@@ -1,3 +1,10 @@
+<a name="1.0.0-alpha.7"></a>
+# 1.0.0-alpha.7 (2019-05-24)
+
+### Breaking
+
+* move to the org scoped package `@deckdeckgo/qrcode`
+
 <a name="v1.0.0-alpha.6"></a>
 # [1.0.0-alpha.6](https://github.com/deckgo/deckdeckgo-qrcode/compare/v1.0.0-alpha.5...v1.0.0-alpha.6) (2019-02-07)